requirements.txt
async-timeout==4.0.2
blinker==1.6.2
click==8.1.6
Flask==2.3.2
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
pip==21.2.4
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
setuptools==58.0.4
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.6
zipp==3.16.2
